refactor(discovery): move AudioEngine to services as a shared singleton
AudioEngine is an application-wide DI singleton consumed by both the
discovery canvas and the unrelated settings route, so it belongs in
src/services/ (reserved for app-wide singletons per CLAUDE.md) rather
than the dna-orb component folder. The cross-boundary import from the
settings route was the tell. No behavior change.

Refs: #373
