+++ src/main.ts
@@ -23,7 +23,6 @@ import { ArtistUnfollowSheet } from './components/artist-unfollow-sheet/artist-u
 import { BottomNavBar } from './components/bottom-nav-bar/bottom-nav-bar'
 import { BottomSheet } from './components/bottom-sheet/bottom-sheet'
 import { CelebrationOverlay } from './components/celebration-overlay/celebration-overlay'
-import { IAudioEngine } from './components/dna-orb/audio-engine'
 import { InlineError } from './components/inline-error/inline-error'
 import { ConcertHighway } from './components/live-highway/concert-highway'
 import { EventCard } from './components/live-highway/event-card'
@@ -59,6 +58,7 @@ import { AuthHook } from './hooks/auth-hook'
 import en from './locales/en/translation.json'
 import ja from './locales/ja/translation.json'
 import { IArtistServiceClient } from './services/artist-service-client'
+import { IAudioEngine } from './services/audio-engine'
 import { IAuthService } from './services/auth-service'
 import { IConcertService } from './services/concert-service'
 import { IErrorBoundaryService } from './services/error-boundary-service'
